Set up checkout from cart

+++ shared/src/commonMain/kotlin/com/vickbt/carrystore/data/mappers/EntityToDomain.kt
@@ -8,7 +8,7 @@ fun ProductEntity.toDomain(): Product {
         id = id.toInt(),
         name = name,
         description = description,
-        price = price.toInt(),
+        price = price.toDouble(),
         currencyCode = currencyCode,
         currencySymbol = currencySymbol,
         quantity = quantity.toInt(),
